Use fixed, full path to test_fixtures.
Using relative paths and full paths based on process.cwd() both cause
tests to fail when executed using the VS Code debugger, because the
working directory is set to the directory of the executed file, but the
tests expect the working directory to be the pwa-kit-runtime directory.
Using fixed full paths relative to the test file solves this problem.