[App Extensibility] Bring back and fix the recently-skipped tests (@W-18120857@) (#2349)
* Bring back the data-cloud tests

For some reason, jest cannot mock js-cookie. It crashed with out-of-memory error.

So I decided to avoid mocking it. Calling `Cookie.get` would return either a string or undefined (if not found). So I adjusted our mocks to allow for undefined value as well.

* Fix header tests

My Account icon actually gets rendered with 'My account' label (note that it's a small-case "a").

* Revert "Bring back the data-cloud tests"

This reverts commit 16978e12f67c01c081f3ec1de092d06f54887278.

* Mock js-cookie in a different way

Such that we won't run into the out-of-memory error.